fix(frontend): disable csp
While the Content Security Policy is a very useful tool to ensure security, the project is in development with no active users. At the moment as a developer, I'd like to focus on the actual product itself. I normally would make the CSP only apply in production, however this causes bugs that only appear in production. A developer's worst nightmare.